refactor(memory): move the sixteen inline test modules to sibling files (#1918)
Fifth crate of the per-crate migration campaign. Fourteen files are a
straight #[path] sibling move, including mutation.rs whose block sat
mid-file with production code and further sibling declarations after it.
context.rs and http.rs each carried two separately named inline modules
and were split by hand into one sibling per module, keeping every module
name (chunk_policy_tests, warning_completeness_tests, body_cap_tests,
keep_alive_tests). The sorted test-name list is byte-identical before
and after (787 lib tests).

context.rs and extract.rs stay above the 1000-line budget but shrank, so
their frozen entries are refreshed; the inline-tests baseline shrinks by
the sixteen velesdb-memory lines (1637 lines paid), leaving 157 files.
