+++ crates/velesdb-memory/src/context.rs
@@ -1079,39 +1079,8 @@ fn saved_by_rule(
 mod media_pipeline_tests;
 
 #[cfg(test)]
-mod chunk_policy_tests {
-    use super::{effective_chunk_policy, MIN_CHUNK_BYTES};
-    use crate::context::estimator::HeuristicEstimator;
-    use crate::context::model::CompilePolicy;
-
-    #[test]
-    fn test_effective_chunk_policy_floors_chunk_size_under_a_tiny_budget() {
-        // A budget of one usable token must NOT drive the chunk ceiling down
-        // toward a byte, which would explode a large fragment into one heap
-        // String per byte (a caller-controlled memory-amplification DoS).
-        let policy = CompilePolicy::default();
-        let effective = effective_chunk_policy(&policy, 1, &HeuristicEstimator);
-        assert!(
-            effective.max_chunk_bytes >= MIN_CHUNK_BYTES,
-            "tiny budget drove chunk size to {} bytes, below the {MIN_CHUNK_BYTES}-byte floor",
-            effective.max_chunk_bytes
-        );
-    }
-
-    #[test]
-    fn test_effective_chunk_policy_floors_a_caller_supplied_tiny_chunk_size() {
-        // A caller cannot bypass the floor by setting a tiny max_chunk_bytes
-        // in the request policy — the same amplification vector otherwise.
-        let mut policy = CompilePolicy::default();
-        policy.chunk.max_chunk_bytes = 1;
-        let effective = effective_chunk_policy(&policy, 1_000, &HeuristicEstimator);
-        assert!(
-            effective.max_chunk_bytes >= MIN_CHUNK_BYTES,
-            "caller max_chunk_bytes=1 bypassed the {MIN_CHUNK_BYTES}-byte floor, got {}",
-            effective.max_chunk_bytes
-        );
-    }
-}
+#[path = "chunk_policy_tests.rs"]
+mod chunk_policy_tests;
 
 /// #1703 DC-4. The published descriptions used to say that checking
 /// `decisions` by hand was "only needed when `warnings` is non-empty". These
@@ -1124,77 +1093,5 @@ mod chunk_policy_tests {
 /// drive a packer into a partial `Preserve` would pin the packer's tuning
 /// instead of the contract.
 #[cfg(test)]
-mod warning_completeness_tests {
-    use super::{warnings_for, WARNING_RELEVANCE_THRESHOLD};
-    use crate::context::model::{ContextAction, ContextDecision, FidelityRisk};
-
-    fn decision(action: ContextAction, relevance: f32, reason: &str) -> ContextDecision {
-        ContextDecision {
-            fragment_id: 1,
-            content_hash: 0,
-            action,
-            rule_id: "test".to_owned(),
-            relevance,
-            risk: FidelityRisk::Medium,
-            reason: reason.to_owned(),
-            memory_id: None,
-            handle: None,
-        }
-    }
-
-    #[test]
-    fn an_empty_warnings_list_is_not_a_clean_bill_of_health() {
-        // Every one of these is a REAL loss for the caller, and not one of
-        // them clears the `Retrieve`-only filter. A caller who trusted the
-        // old shortcut shipped all four believing nothing was cut.
-        let decisions = vec![
-            decision(
-                ContextAction::Preserve,
-                0.9,
-                "packed 2/9 chunks — the rest did not fit",
-            ),
-            decision(ContextAction::Abstract, 0.9, "summarised"),
-            decision(
-                ContextAction::Drop,
-                0.9,
-                "duplicate — image survives through it; this fragment's differing caption does not",
-            ),
-            decision(
-                ContextAction::Drop,
-                0.9,
-                "duplicate — but that twin was not fully emitted — recover via the handle",
-            ),
-        ];
-
-        assert!(
-            warnings_for(&decisions).is_empty(),
-            "the filter is Retrieve-only, so these four losses must NOT warn — \
-             if this ever starts warning, the published descriptions that now \
-             say 'an empty warnings is not a clean bill of health' become the \
-             stale ones and must be revisited"
-        );
-    }
-
-    #[test]
-    fn the_relevance_floor_silences_a_retrieve_that_is_still_a_loss() {
-        // The second half of the same lie: even the ONE action that can warn
-        // stays silent below the floor.
-        let below = decision(
-            ContextAction::Retrieve,
-            WARNING_RELEVANCE_THRESHOLD - 0.01,
-            "externalized behind a handle",
-        );
-        assert!(warnings_for(std::slice::from_ref(&below)).is_empty());
-
-        let at_floor = decision(
-            ContextAction::Retrieve,
-            WARNING_RELEVANCE_THRESHOLD,
-            "externalized behind a handle",
-        );
-        assert_eq!(
-            warnings_for(std::slice::from_ref(&at_floor)).len(),
-            1,
-            "the floor is inclusive — at exactly the threshold it must warn"
-        );
-    }
-}
+#[path = "warning_completeness_tests.rs"]
+mod warning_completeness_tests;
