+++ scripts/inline-tests-baseline.txt
@@ -79,22 +79,6 @@ crates/velesdb-core/src/velesql/error.rs	67
 crates/velesdb-core/src/velesql/parser/helpers.rs	144
 crates/velesdb-core/src/velesql/parser/hints.rs	49
 crates/velesdb-core/src/wire/vrb1.rs	119
-crates/velesdb-memory/src/context.rs	109
-crates/velesdb-memory/src/daemon_commands.rs	133
-crates/velesdb-memory/src/daemon_serve.rs	19
-crates/velesdb-memory/src/extract.rs	182
-crates/velesdb-memory/src/http.rs	63
-crates/velesdb-memory/src/http/session_limit.rs	307
-crates/velesdb-memory/src/http_retry.rs	160
-crates/velesdb-memory/src/id.rs	80
-crates/velesdb-memory/src/logging.rs	74
-crates/velesdb-memory/src/mcp/migration_tools.rs	55
-crates/velesdb-memory/src/migration/diagnosis/capabilities.rs	18
-crates/velesdb-memory/src/migration/filesystem.rs	21
-crates/velesdb-memory/src/mutation.rs	144
-crates/velesdb-memory/src/mutation/journal/format.rs	32
-crates/velesdb-memory/src/reinforce.rs	167
-crates/velesdb-memory/src/tls.rs	73
 crates/velesdb-migrate/src/config.rs	202
 crates/velesdb-migrate/src/connectors/chromadb.rs	47
 crates/velesdb-migrate/src/connectors/common.rs	271
